// src/app/models/user.ts
export interface IUser {
  userId: number;
  userName: string;
  userPassword: string;
  userEmail: string;
}

export interface IEmployee {
  id: number;
  fName: string;
  lName: string;
  technology: string;
  designation: string;
  skills: string;
  coreExpertise: string;
  isCertification: boolean;
  certificationName: string;
  compnay: string;
  about: string;
}